♻️ [Refactor/room] determine participant count by location input count on getting room detail API (#107)
* :recycle: refactor: determine participant count by location input count

* :white_check_mark: test: fix test response

+++ kok-core/src/main/java/com/kok/kokcore/room/usecase/GetRoomUseCase.java
@@ -13,5 +13,5 @@ public interface GetRoomUseCase {
 
     Member getParticipant(String roomId, String memberId);
 
-    int getParticipantsCount(String roomId);
+    long getParticipantsCount(String roomId);
 }
